data/airflow/requirements.txt
# ============================================
# Airflow Requirements - Enhanced Version
# ============================================

# Core Airflow Providers
apache-airflow-providers-mysql>=5.0.0
apache-airflow-providers-postgres>=5.8.0
apache-airflow-providers-http>=4.5.0
apache-airflow-providers-amazon>=8.0.0
apache-airflow-providers-google>=10.0.0
apache-airflow-providers-microsoft-azure>=7.0.0

# Data Quality & Validation
great-expectations>=0.18.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic-extra-types>=2.3.0
email-validator>=2.1.0
pandera>=0.17.0

# Database
psycopg2-binary>=2.9.7
asyncpg>=0.29.0  # Async PostgreSQL driver (faster)
sqlalchemy>=2.0.23
alembic>=1.13.0

# HTTP & API Clients
requests>=2.31.0
httpx>=0.25.0  # Modern async HTTP client
httpx[http2]>=0.25.0
aiohttp>=3.9.0  # Alternative async HTTP
httpcore>=1.0.0

# Cloud Storage
boto3>=1.34.0  # AWS S3
azure-storage-blob>=12.19.0  # Azure Blob Storage
google-cloud-storage>=2.14.0  # GCP Cloud Storage
s3fs>=2023.12.0  # S3 filesystem interface

# Resiliencia & Circuit Breakers
tenacity>=8.2.2  # Retry logic mejorado
backoff>=2.2.1  # Alternative retry decorator
pybreaker>=1.0.1  # Advanced circuit breaker
circuitbreaker>=2.0.0  # Circuit breaker pattern

# Rate Limiting & Throttling
aiolimiter>=1.1.0  # Async rate limiter
asyncio-throttle>=1.0.2
limits>=3.6.0  # Generic rate limiting
slowapi>=0.1.9  # Rate limiting for Flask/FastAPI

# Logging & Observabilidad
structlog>=23.2.0  # Structured logging
loguru>=0.7.2  # Modern logging
python-json-logger>=2.0.7

# OpenTelemetry - Distributed Tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-flask>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-exporter-otlp>=1.21.0

# Métricas & Monitoring
prometheus-client>=0.19.0
psutil>=5.9.8  # System monitoring

# Caching
redis>=5.0.1
aioredis>=2.0.1  # Async Redis
cachetools>=5.3.2
diskcache>=5.6.3
aiocache>=0.12.2  # Async cache

# OCR & Document Processing
pytesseract>=0.3.10
Pillow>=10.1.0
google-cloud-vision>=3.4.5
PyPDF2>=3.0.1
pdf2image>=1.16.3
opencv-python>=4.8.1.78
PyMuPDF>=1.23.8  # PDF compression

# Data Processing
pandas>=2.1.3
numpy>=1.26.2
polars>=0.19.19  # Fast DataFrame alternative
pyarrow>=14.0.1  # Columnar data processing
openpyxl>=3.1.2  # Excel support

# Machine Learning
scikit-learn>=1.3.2
sentence-transformers>=2.2.2
transformers>=4.36.0  # Optional: abstractive summarization
torch>=2.1.0  # Required by transformers

# NLP & Text Processing
spacy>=3.7.2  # NLP and NER
vaderSentiment>=3.3.2  # Sentiment analysis
textblob>=0.17.1

# Translation
googletrans==4.0.0-rc1  # Google Translate
deepl>=1.15.0  # DeepL (better quality)

# Vector Databases
pinecone-client>=2.2.4  # Pinecone
qdrant-client>=1.6.9  # Qdrant
elasticsearch>=8.11.0  # Search indexing

# Serialización Rápida
orjson>=3.9.10  # Ultra-fast JSON serializer
ujson>=5.9.0  # Fast JSON alternative
msgpack>=1.0.7  # Binary serialization

# Security
cryptography>=41.0.3
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0  # JWT handling
passlib[bcrypt]>=1.7.4  # Password hashing
authlib>=1.2.1  # OAuth/OpenID Connect
bleach>=6.1.0  # HTML sanitization
markupsafe>=2.1.3
defusedxml>=0.7.1  # Safe XML parsing

# API & Web Framework
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.104.1  # Modern API framework (optional)
uvicorn[standard]>=0.24.0  # ASGI server
starlette>=0.27.0

# Utilities
python-dateutil>=2.8.2
pendulum>=3.0.0  # Date/time handling
arrow>=1.3.0  # Modern date handling
pytz>=2023.3.post1
python-dotenv>=1.0.0  # Environment variables
dynaconf>=3.2.4  # Dynamic configuration
click>=8.1.7  # CLI framework
rich>=13.7.0  # Rich terminal formatting
tqdm>=4.66.1  # Progress bars
humanize>=4.8.0  # Human-readable formatting

# Kubernetes & Infrastructure
kubernetes>=28.1.0  # Kubernetes client
pyyaml>=6.0.1  # YAML handling

# Message Queue
kafka-python>=2.0.2
celery>=5.3.4  # Task queue
celery[redis]>=5.3.4

# Async Utilities
asyncio-timeout>=4.0.3
aiofiles>=23.2.1  # Async file I/O

# Data Validation & Transformation
marshmallow>=3.20.1  # Serialization/deserialization
cattrs>=23.2.0  # Structured data transformation
jsonschema>=4.20.0  # JSON schema validation
voluptuous>=0.13.1  # Declarative validation

# Type Safety (Runtime)
typeguard>=4.1.0  # Runtime type checking

# Performance Profiling
pyinstrument>=5.5.0  # Performance profiling
memory-profiler>=0.61.0  # Memory profiling

# Image Processing
scikit-image>=0.22.0  # Advanced image processing

# Web Scraping
beautifulsoup4>=4.12.2  # HTML parsing for web scraping
lxml>=4.9.3  # Fast XML/HTML parser

# Reporting
reportlab>=4.0.7  # PDF reports
